Revert "Сделал оба класса, пока только их"
This reverts commit 15d05972
